+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* 封装`dio`，实现基本的`get、post、patch、put、delete`请求。请求接支持口传入自定义`Codable`模型，实现自动json解析的功能